+++ town-api/src/main/java/po/UserReadNoticeInfoOrBuilder.java
@@ -34,7 +34,7 @@ public interface UserReadNoticeInfoOrBuilder extends
    * 阅读用户电话
    * </pre>
    *
-   * <code>optional int32 userTel = 2;</code>
+   * <code>optional int64 userTel = 2;</code>
    * @return Whether the userTel field is set.
    */
   boolean hasUserTel();
@@ -43,10 +43,10 @@ public interface UserReadNoticeInfoOrBuilder extends
    * 阅读用户电话
    * </pre>
    *
-   * <code>optional int32 userTel = 2;</code>
+   * <code>optional int64 userTel = 2;</code>
    * @return The userTel.
    */
-  int getUserTel();
+  long getUserTel();
 
   /**
    * <pre>
